app/requirements.txt
django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==3.3
drf-yasg==1.20.0
